+++ lib/mongoid/shell/errors.rb
@@ -1,2 +1,3 @@
 require 'mongoid/shell/errors/base'
 require 'mongoid/shell/errors/missing_session_error'
+require 'mongoid/shell/errors/session_not_connected_error'