+++ lib/mongoid/shell/errors/session_not_connected_error.rb
@@ -1,0 +1,11 @@
+module Mongoid
+  module Shell
+    module Errors
+      class SessionNotConnectedError < Mongoid::Shell::Errors::Base
+        def initialize
+          super(compose_message("session_not_connected"))
+        end
+      end
+    end
+  end
+end