+++ lib/mongoid/shell/properties/host.rb
@@ -5,7 +5,9 @@ module Mongoid
 
         # database host
         def host
-          session.cluster.nodes.first.address
+          node = session.cluster.nodes.first
+          raise Mongoid::Shell::Errors::SessionNotConnectedError unless node
+          node.address == "localhost:27017" ? nil : node.address
         end
 
       end