+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+Next Release
+============
+
+* A `Mongoid::Shell::Errors::SessionNotConnectedError` error will be raised when a session is not connected - [@dblock](https://github.com/dblock).
+
 0.1.0 (1/27/2013)
 ==================